Refactored the dynamic element to use a hook for part of it
